# api_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solar-Eye API",
    "version": "1.0.0",
    "description": "AI 기반 태양광 패널 결함 탐지 및 유지보수 솔루션 API 명세서"
  },
  "servers": [
    {
      "url": "https://api.solar-eye.com",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/api/v1/weather/current": {
      "get": {
        "tags": ["Weather"],
        "summary": "현재 날씨 정보 조회",
        "description": "기상청 초단기실황 API를 통해 현재 날씨 정보를 조회합니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "패널 ID (패널의 위치 기준 날씨 조회)"
          },
          {
            "name": "nx",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "기상청 격자 X 좌표 (panel_id 미지정 시 필수)"
          },
          {
            "name": "ny",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "기상청 격자 Y 좌표 (panel_id 미지정 시 필수)"
          }
        ],
        "responses": {
          "200": {
            "description": "현재 날씨 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/CurrentWeather" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weather/forecast": {
      "get": {
        "tags": ["Weather"],
        "summary": "단기 예보 조회",
        "description": "12시간/3일 기상 예보 정보를 조회합니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "nx",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "ny",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "hours",
            "in": "query",
            "schema": { "type": "integer", "default": 12, "enum": [12, 24, 48, 72] },
            "description": "예보 시간 범위"
          }
        ],
        "responses": {
          "200": {
            "description": "단기 예보 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "location": {
                          "type": "object",
                          "properties": {
                            "nx": { "type": "integer", "example": 60 },
                            "ny": { "type": "integer", "example": 127 }
                          }
                        },
                        "forecasts": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ForecastItem" }
                        },
                        "fetched_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weather/status": {
      "get": {
        "tags": ["Weather"],
        "summary": "우천 여부 확인",
        "description": "현재 비/눈 여부를 확인하여 알림 중지 판단에 활용합니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "우천 상태",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "is_rainy": { "type": "boolean", "example": false, "description": "우천 여부 (true=비/눈)" },
                        "precipitation_type": { 
                          "type": "string", 
                          "enum": ["none", "rain", "rain_snow", "snow", "shower"],
                          "example": "none",
                          "description": "강수형태"
                        },
                        "soiling_alert_enabled": { "type": "boolean", "example": true, "description": "오염 알림 활성화 여부" },
                        "checked_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID Token"
      }
    },
    "schemas": {
...
      "CurrentWeather": {
        "type": "object",
        "properties": {
          "location": {
            "type": "object",
            "properties": {
              "nx": { "type": "integer", "example": 60, "description": "기상청 격자 X 좌표" },
              "ny": { "type": "integer", "example": 127, "description": "기상청 격자 Y 좌표" }
            }
          },
          "temperature": { "type": "number", "example": 5.2, "description": "현재 기온 (°C)" },
          "humidity": { "type": "integer", "example": 45, "description": "습도 (%)" },
          "precipitation_type": { 
            "type": "string", 
            "enum": ["none", "rain", "rain_snow", "snow", "shower"],
            "example": "none",
            "description": "강수형태 (PTY: 0=없음, 1=비, 2=비/눈, 3=눈, 4=소나기)"
          },
          "sky_condition": {
            "type": "string",
            "enum": ["clear", "partly_cloudy", "cloudy", "overcast"],
            "example": "clear",
            "description": "하늘 상태 (SKY: 1=맑음, 3=구름많음, 4=흐림)"
          },
          "wind_speed": { "type": "number", "example": 2.5, "description": "풍속 (m/s)" },
          "precipitation_1h": { "type": "number", "example": 0, "description": "1시간 강수량 (mm)" },
          "is_rainy": { "type": "boolean", "example": false, "description": "우천 여부" },
          "condition_text": { "type": "string", "example": "맑음", "description": "날씨 상태 텍스트" },
          "condition_icon": { "type": "string", "example": "sunny", "description": "날씨 아이콘 코드" },
          "observed_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:00:00Z" },
          "fetched_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:05:00Z" }
        }
      },
      "ForecastItem": {
        "type": "object",
        "properties": {
          "forecast_time": { "type": "string", "format": "date-time", "example": "2026-01-16T18:00:00Z" },
          "temperature": { "type": "number", "example": 3.0 },
          "humidity": { "type": "integer", "example": 50 },
          "precipitation_probability": { "type": "integer", "example": 10, "description": "강수확률 (%)" },
          "precipitation_type": { "type": "string", "enum": ["none", "rain", "rain_snow", "snow", "shower"] },
          "sky_condition": { "type": "string", "enum": ["clear", "partly_cloudy", "cloudy", "overcast"] },
          "condition_text": { "type": "string", "example": "구름많음" },
          "condition_icon": { "type": "string", "example": "partly_cloudy" }
        }
      }
    }
  }
}